pyproject.toml
[project]
name = "neonbot"
version = "2.3.1"
description = ""
authors = [
    {name = "NeonSpectrum",email = "youngskymann@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "environs (>=14.1.1,<15.0.0)",
    "motor (>=3.7.0,<4.0.0)",
    "python-i18n (>=0.3.9,<0.4.0)",
    "dnspython (>=2.7.0,<3.0.0)",
    "pynacl (>=1.5.0,<2.0.0)",
    "yt-dlp (>=2025.3.31)",
    "pytz (>=2025.2,<2026.0)",
    "envparse (>=0.2.0,<0.3.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "fuzzywuzzy (>=0.18.0,<0.19.0)",
    "python-levenshtein (>=0.27.1,<0.28.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "beanie (>=1.29.0,<2.0.0)",
    "openai (>=1.75.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "validators (>=0.34.0,<0.35.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "markdown (>=3.8,<4.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "ytmusicapi (>=1.10.3,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "discord-py[voice] (>=2.5.2,<3.0.0)",
    "jikanpy-v4 (>=1.0.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true